+++ integration/helpers/vite.ts
@@ -1,11 +1,10 @@
 import { spawn, spawnSync, type ChildProcess } from "node:child_process";
-import path from "pathe";
-import fs from "node:fs/promises";
-import type { Readable } from "node:stream";
-import url from "node:url";
+import { cp, mkdir, readFile, writeFile } from "node:fs/promises";
 import { createRequire } from "node:module";
 import { platform } from "node:os";
-import fse from "fs-extra";
+import type { Readable } from "node:stream";
+import url from "node:url";
+import path from "pathe";
 import stripIndent from "strip-indent";
 import waitOn from "wait-on";
 import getPort from "get-port";
@@ -189,6 +188,8 @@ export const EXPRESS_SERVER = (args: {
 type FrameworkModeViteMajorTemplateName =
   | "vite-5-template"
   | "vite-6-template"
+  | "vite-7-beta-template"
+  | "vite-plugin-cloudflare-template"
   | "vite-rolldown-template";
 
 type FrameworkModeRscTemplateName = "rsc-parcel-framework";
@@ -208,6 +209,7 @@ export type TemplateName =
 export const viteMajorTemplates = [
   { templateName: "vite-5-template", templateDisplayName: "Vite 5" },
   { templateName: "vite-6-template", templateDisplayName: "Vite 6" },
+  { templateName: "vite-7-beta-template", templateDisplayName: "Vite 7 Beta" },
   {
     templateName: "vite-rolldown-template",
     templateDisplayName: "Vite Rolldown",
@@ -231,18 +233,18 @@ export async function createProject(
 ) {
   let projectName = `rr-${Math.random().toString(32).slice(2)}`;
   let projectDir = path.join(TMP_DIR, projectName);
-  await fse.ensureDir(projectDir);
+  await mkdir(projectDir, { recursive: true });
 
   // base template
   let templateDir = path.resolve(__dirname, templateName);
-  await fse.copy(templateDir, projectDir, { errorOnExist: true });
+  await cp(templateDir, projectDir, { errorOnExist: true, recursive: true });
 
   // user-defined files
   await Promise.all(
     Object.entries(files).map(async ([filename, contents]) => {
       let filepath = path.join(projectDir, filename);
-      await fse.ensureDir(path.dirname(filepath));
-      await fse.writeFile(filepath, stripIndent(contents));
+      await mkdir(path.dirname(filepath), { recursive: true });
+      await writeFile(filepath, stripIndent(contents));
     })
   );
 
@@ -251,7 +253,7 @@ export async function createProject(
 
 // Avoid "Warning: The 'NO_COLOR' env is ignored due to the 'FORCE_COLOR' env
 // being set" in vite-ecosystem-ci which breaks empty stderr assertions. To fix
-// this we always ensure that only NO_COLOR is set after spreading process.env.
+// this, we always ensure that only NO_COLOR is set after spreading process.env.
 const colorEnv = {
   FORCE_COLOR: undefined,
   NO_COLOR: "1",
@@ -512,11 +514,11 @@ export function createEditor(projectDir: string) {
     transform: (contents: string) => string
   ) {
     let filepath = path.join(projectDir, file);
-    let contents = await fs.readFile(filepath, "utf8");
-    await fs.writeFile(filepath, transform(contents), "utf8");
+    let contents = await readFile(filepath, "utf8");
+    await writeFile(filepath, transform(contents), "utf8");
 
     return async function revert() {
-      await fs.writeFile(filepath, contents, "utf8");
+      await writeFile(filepath, contents, "utf8");
     };
   };
 }