+++ packages/react-router-dev/vite/plugin.ts
@@ -1,11 +1,19 @@
 // We can only import types from Vite at the top level since we're in a CJS
-// context but want to use Vite's ESM build to avoid deprecation warnings
+// context but want to use Vite's ESM build since Vite 7+ is ESM only
 import type * as Vite from "vite";
 import { type BinaryLike, createHash } from "node:crypto";
-import * as fs from "node:fs";
+import { existsSync, readFileSync, readdirSync, rmSync } from "node:fs";
+import {
+  cp,
+  mkdir,
+  readdir,
+  readFile,
+  rename,
+  rm,
+  writeFile,
+} from "node:fs/promises";
 import * as path from "node:path";
 import * as url from "node:url";
-import * as fse from "fs-extra";
 import * as babel from "@babel/core";
 import {
   unstable_setDevServerHooks as setDevServerHooks,
@@ -23,6 +31,7 @@ import {
   init as initEsModuleLexer,
   parse as esModuleLexer,
 } from "es-module-lexer";
+import { escapePath as escapePathAsGlob } from "tinyglobby";
 import pick from "lodash/pick";
 import jsesc from "jsesc";
 import colors from "picocolors";
@@ -136,7 +145,7 @@ const CLIENT_ROUTE_EXPORTS = [
 
 /** This is used to manage a build optimization to remove unused route exports
 from the client build output. This is important in cases where custom route
-exports are only ever used on the server. Without this optimization we can't
+exports are only ever used on the server. Without this optimization, we can't
 tree-shake any unused custom exports because routes are entry points. */
 const BUILD_CLIENT_ROUTE_QUERY_STRING = "?__react-router-build-client-route";
 
@@ -392,8 +401,8 @@ function dedupe<T>(array: T[]): T[] {
 }
 
 const writeFileSafe = async (file: string, contents: string): Promise<void> => {
-  await fse.ensureDir(path.dirname(file));
-  await fse.writeFile(file, contents);
+  await mkdir(path.dirname(file), { recursive: true });
+  await writeFile(file, contents);
 };
 
 const getExportNames = (code: string): string[] => {
@@ -447,7 +456,7 @@ const compileRouteFile = async (
 
   let [id, code] = await Promise.all([
     resolveId(),
-    readRouteFile?.() ?? fse.readFile(routePath, "utf-8"),
+    readRouteFile?.() ?? readFile(routePath, "utf-8"),
     // pluginContainer.transform(...) fails if we don't do this first:
     moduleGraph.ensureEntryFromUrl(url, ssr),
   ]);
@@ -551,9 +560,9 @@ let defaultEntriesDir = path.resolve(
   "config",
   "defaults"
 );
-let defaultEntries = fse
-  .readdirSync(defaultEntriesDir)
-  .map((filename) => path.join(defaultEntriesDir, filename));
+let defaultEntries = readdirSync(defaultEntriesDir).map((filename) =>
+  path.join(defaultEntriesDir, filename)
+);
 invariant(defaultEntries.length > 0, "No default entries found");
 
 type MaybePromise<T> = T | Promise<T>;
@@ -597,7 +606,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
   // change or route file addition/removal.
   let ctx: ReactRouterPluginContext;
 
-  /** Mutates `ctx` as a side-effect */
+  /** Mutates `ctx` as a side effect */
   let updatePluginContext = async (): Promise<void> => {
     let reactRouterConfig: ResolvedReactRouterConfig;
     let reactRouterConfigResult = await reactRouterConfigLoader.getConfig();
@@ -614,7 +623,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
 
     // This `injectedPluginContext` logic is so we can support injecting an
     // already-resolved plugin context into the build in case we want to re-use
-    // any resolved values. Currently this is used so that we can re-use the
+    // any resolved values. Currently, this is used so that we can re-use the
     // `buildManifest` object across multiple builds without having to
     // re-execute the `serverBundles` function.
     let injectedPluginContext =
@@ -769,7 +778,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
   };
 
   let loadViteManifest = async (directory: string) => {
-    let manifestContents = await fse.readFile(
+    let manifestContents = await readFile(
       path.resolve(directory, ".vite", "manifest.json"),
       "utf-8"
     );
@@ -803,7 +812,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
   let generateSriManifest = async (ctx: ReactRouterPluginContext) => {
     let clientBuildDirectory = getClientBuildDirectory(ctx.reactRouterConfig);
     // walk the client build directory and generate SRI hashes for all .js files
-    let entries = fs.readdirSync(clientBuildDirectory, {
+    let entries = readdirSync(clientBuildDirectory, {
       withFileTypes: true,
       recursive: true,
     });
@@ -817,7 +826,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
 
         let contents;
         try {
-          contents = await fse.readFile(
+          contents = await readFile(
             path.join(entryNormalizedPath, entry.name),
             "utf-8"
           );
@@ -1145,6 +1154,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
 
         // Ensure sync import of Vite works after async preload
         let vite = getVite();
+        let viteMajorVersion = parseInt(vite.version.split(".")[0], 10);
 
         viteUserConfig = _viteUserConfig;
         viteConfigEnv = _viteConfigEnv;
@@ -1192,7 +1202,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
           vite.loadEnv(
             viteConfigEnv.mode,
             viteUserConfig.envDir ?? ctx.rootDirectory,
-            // We override default prefix of "VITE_" with a blank string since
+            // We override the default prefix of "VITE_" with a blank string since
             // we're targeting the server, so we want to load all environment
             // variables, not just those explicitly marked for the client
             ""
@@ -1232,7 +1242,11 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
                   ...Object.values(ctx.reactRouterConfig.routes).map((route) =>
                     resolveRelativeRouteFilePath(route, ctx.reactRouterConfig)
                   ),
-                ]
+                ].map((entry) =>
+                  // In Vite 7, the `optimizeDeps.entries` option only accepts glob patterns.
+                  // In prior versions, absolute file paths were treated differently.
+                  viteMajorVersion >= 7 ? escapePathAsGlob(entry) : entry
+                )
               : [],
             include: [
               // Pre-bundle React dependencies to avoid React duplicates,
@@ -1276,7 +1290,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
           },
           base: viteUserConfig.base,
 
-          // When consumer provides an allow list for files that can be read by
+          // When consumer provides an allowlist for files that can be read by
           // the server, ensure that the default entry files are included.
           // If we don't do this and a default entry file is used, the server
           // will throw an error that the file is not allowed to be read.
@@ -1469,7 +1483,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
                   plugin.name !== "react-router:route-exports" &&
                   plugin.name !== "react-router:hmr-updates"
               )
-              // Remove server hooks to avoid conflicts with main dev server
+              // Remove server hooks to avoid conflicts with the main dev server
               .map((plugin) => ({
                 ...plugin,
                 configureServer: undefined,
@@ -1576,7 +1590,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
         if (ctx.reactRouterConfig.future.unstable_viteEnvironmentApi) {
           viteDevServer.middlewares.use(async (req, res, next) => {
             let [reqPathname, reqSearch] = (req.url ?? "").split("?");
-            if (reqPathname === `${ctx.publicPath}@react-router/critical.css`) {
+            if (reqPathname.endsWith("/@react-router/critical.css")) {
               let pathname = new URLSearchParams(reqSearch).get("pathname");
               if (!pathname) {
                 return next("No pathname provided");
@@ -1692,15 +1706,15 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
             let dest = path.join(clientBuildDirectory, ssrAssetPath);
 
             if (!userSsrEmitAssets) {
-              if (!fse.existsSync(dest)) {
-                await fse.move(src, dest);
+              if (!existsSync(dest)) {
+                await rename(src, dest);
                 movedAssetPaths.push(dest);
               } else {
-                await fse.remove(src);
+                await rm(src, { force: true, recursive: true });
                 removedAssetPaths.push(dest);
               }
-            } else if (!fse.existsSync(dest)) {
-              await fse.copy(src, dest);
+            } else if (!existsSync(dest)) {
+              await cp(src, dest, { recursive: true });
               copiedAssetPaths.push(dest);
             }
           }
@@ -1714,7 +1728,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
             await Promise.all(
               ssrCssPaths.map(async (cssPath) => {
                 let src = path.join(serverBuildDirectory, cssPath);
-                await fse.remove(src);
+                await rm(src, { force: true, recursive: true });
                 removedAssetPaths.push(src);
               })
             );
@@ -1728,9 +1742,9 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
           await Promise.all(
             Array.from(cleanedAssetDirs).map(async (dir) => {
               try {
-                const files = await fse.readdir(dir);
+                const files = await readdir(dir, { recursive: true });
                 if (files.length === 0) {
-                  await fse.remove(dir);
+                  await rm(dir, { force: true, recursive: true });
                 }
               } catch {}
             })
@@ -1823,7 +1837,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
                 "due to ssr:false",
               ].join(" ")
             );
-            fse.removeSync(serverBuildDirectory);
+            rmSync(serverBuildDirectory, { force: true, recursive: true });
           }
         },
       },
@@ -1842,7 +1856,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
       // primarily ensures code is never duplicated across a route module and
       // its chunks. If we didn't have this plugin, any app that explicitly
       // imports a route module would result in duplicate code since the app
-      // would contain code for both the unprocessed route module as well as its
+      // would contain code for both the unprocessed route module and its
       // individual chunks. This is because, since they have different module
       // IDs, they are treated as completely separate modules even though they
       // all reference the same underlying file. This plugin addresses this by
@@ -1950,7 +1964,7 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
         // Routes aren't chunked on the server
         if (options?.ssr) return;
 
-        // Ignore anything that isn't marked as a route chunk
+        // Ignore anything not marked as a route chunk
         if (!isRouteChunkModuleId(id)) return;
 
         invariant(
@@ -2247,11 +2261,8 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
 
         return [
           "const exports = {}",
-          await fse.readFile(reactRefreshRuntimePath, "utf8"),
-          await fse.readFile(
-            require.resolve("./static/refresh-utils.cjs"),
-            "utf8"
-          ),
+          await readFile(reactRefreshRuntimePath, "utf8"),
+          await readFile(require.resolve("./static/refresh-utils.cjs"), "utf8"),
           "export default exports",
         ].join("\n");
       },
@@ -2353,10 +2364,10 @@ export const reactRouterVitePlugin: ReactRouterVitePlugin = () => {
     {
       name: "react-router-server-change-trigger-client-hmr",
       // This hook is only available in Vite v6+ so this is a no-op in v5.
-      // Previously the server and client modules were shared in a single module
+      // Previously, the server and client modules were shared in a single module
       // graph. This meant that changes to server code automatically resulted in
-      // client HMR updates. In Vite v6+ these module graphs are separate from
-      // each other so we need to manually trigger client HMR updates if server
+      // client HMR updates. In Vite v6+, these module graphs are separate from
+      // each other, so we need to manually trigger client HMR updates if server
       // code has changed.
       hotUpdate(this, { server, modules }) {
         if (this.environment.name !== "ssr" && modules.length <= 0) {
@@ -2619,7 +2630,7 @@ async function handleSpaMode(
   let html = await response.text();
 
   // If the user prerendered `/`, then we write this out to a separate file
-  // they can serve.  Otherwise it can be the main entry point.
+  // they can serve. Otherwise, it can be the main entry point.
   let isPrerenderSpaFallback = build.prerender.includes("/");
   let filename = isPrerenderSpaFallback ? "__spa-fallback.html" : "index.html";
   if (response.status !== 200) {
@@ -2649,7 +2660,7 @@ async function handleSpaMode(
   }
 
   // Write out the HTML file for the SPA
-  await fse.writeFile(path.join(clientBuildDirectory, filename), html);
+  await writeFile(path.join(clientBuildDirectory, filename), html);
   let prettyDir = path.relative(process.cwd(), clientBuildDirectory);
   let prettyPath = path.join(prettyDir, filename);
   if (build.prerender.length > 0) {
@@ -2826,8 +2837,8 @@ async function prerenderData(
   // Write out the .data file
   let outdir = path.relative(process.cwd(), clientBuildDirectory);
   let outfile = path.join(outdir, ...normalizedPath.split("/"));
-  await fse.ensureDir(path.dirname(outfile));
-  await fse.outputFile(outfile, data);
+  await mkdir(path.dirname(outfile), { recursive: true });
+  await writeFile(outfile, data);
   viteConfig.logger.info(
     `Prerender (data): ${prerenderPath} -> ${colors.bold(outfile)}`
   );
@@ -2855,7 +2866,7 @@ async function prerenderRoute(
   if (redirectStatusCodes.has(response.status)) {
     // This isn't ideal but gets the job done as a fallback if the user can't
     // implement proper redirects via .htaccess or something else.  This is the
-    // approach used by Astro as well so there's some precedent.
+    // approach used by Astro as well, so there's some precedent.
     // https://github.com/withastro/roadmap/issues/466
     // https://github.com/withastro/astro/blob/main/packages/astro/src/core/routing/3xx.ts
     let location = response.headers.get("Location");
@@ -2885,8 +2896,8 @@ async function prerenderRoute(
   // Write out the HTML file
   let outdir = path.relative(process.cwd(), clientBuildDirectory);
   let outfile = path.join(outdir, ...normalizedPath.split("/"), "index.html");
-  await fse.ensureDir(path.dirname(outfile));
-  await fse.outputFile(outfile, html);
+  await mkdir(path.dirname(outfile), { recursive: true });
+  await writeFile(outfile, html);
   viteConfig.logger.info(
     `Prerender (html): ${prerenderPath} -> ${colors.bold(outfile)}`
   );
@@ -2918,8 +2929,8 @@ async function prerenderResourceRoute(
   // Write out the resource route file
   let outdir = path.relative(process.cwd(), clientBuildDirectory);
   let outfile = path.join(outdir, ...normalizedPath.split("/"));
-  await fse.ensureDir(path.dirname(outfile));
-  await fse.outputFile(outfile, content);
+  await mkdir(path.dirname(outfile), { recursive: true });
+  await writeFile(outfile, content);
   viteConfig.logger.info(
     `Prerender (resource): ${prerenderPath} -> ${colors.bold(outfile)}`
   );
@@ -3310,7 +3321,7 @@ export async function cleanBuildDirectory(
   };
 
   if (viteConfig.build.emptyOutDir ?? isWithinRoot()) {
-    await fse.remove(buildDirectory);
+    await rm(buildDirectory, { force: true, recursive: true });
   }
 }
 
@@ -3327,19 +3338,19 @@ export async function cleanViteManifests(
   );
   await Promise.all(
     viteManifestPaths.map(async (viteManifestPath) => {
-      let manifestExists = await fse.pathExists(viteManifestPath);
+      let manifestExists = existsSync(viteManifestPath);
       if (!manifestExists) return;
 
-      // Delete original Vite manifest file if consumer doesn't want it
+      // Delete the original Vite manifest file if consumer doesn't want it
       if (!ctx.viteManifestEnabled) {
-        await fse.remove(viteManifestPath);
+        await rm(viteManifestPath, { force: true, recursive: true });
       }
 
       // Remove .vite dir if it's now empty
       let viteDir = path.dirname(viteManifestPath);
-      let viteDirFiles = await fse.readdir(viteDir);
+      let viteDirFiles = await readdir(viteDir, { recursive: true });
       if (viteDirFiles.length === 0) {
-        await fse.remove(viteDir);
+        await rm(viteDir, { force: true, recursive: true });
       }
     })
   );
@@ -3527,12 +3538,12 @@ export async function getEnvironmentOptionsResolvers(
       },
       build: {
         // We move SSR-only assets to client assets. Note that the
-        // SSR build can also emit code-split JS files (e.g. by
+        // SSR build can also emit code-split JS files (e.g., by
         // dynamic import) under the same assets directory
         // regardless of "ssrEmitAssets" option, so we also need to
-        // keep these JS files have to be kept as-is.
+        // keep these JS files to be kept as-is.
         ssrEmitAssets: true,
-        copyPublicDir: false, // Assets in the public directory are only used by the client
+        copyPublicDir: false, // The client only uses assets in the public directory
         rollupOptions: {
           input:
             (ctx.reactRouterConfig.future.unstable_viteEnvironmentApi
@@ -3565,7 +3576,7 @@ export async function getEnvironmentOptionsResolvers(
                   let isRootRoute =
                     route.file === ctx.reactRouterConfig.routes.root.file;
 
-                  let code = fse.readFileSync(routeFilePath, "utf-8");
+                  let code = readFileSync(routeFilePath, "utf-8");
 
                   return [
                     `${routeFilePath}${BUILD_CLIENT_ROUTE_QUERY_STRING}`,